u128 => BigDecimal

+++ plasma/src/models/tx.rs
@@ -6,8 +6,8 @@ use super::TxSignature;
 pub struct TransferTx {
     pub from:               u32,
     pub to:                 u32,
-    pub amount:             u128,
-    pub fee:                u128,
+    pub amount:             BigDecimal,
+    pub fee:                BigDecimal,
     pub nonce:              u32,
     pub good_until_block:   u32,
     pub signature:          TxSignature,
@@ -23,11 +23,11 @@ impl TransferTx {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct DepositTx{
     pub account:            u32,
-    pub amount:             u128,
+    pub amount:             BigDecimal,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct ExitTx{
     pub account:            u32,
-    pub amount:             u128,
+    pub amount:             BigDecimal,
 }
